code is fixed
