+++ .gitignore
@@ -1,5 +1,9 @@
 # Ignore Gradle project-specific cache directory
 .gradle
-
+.idea
 # Ignore Gradle build output directory
 build
+gradle
+gradlew
+gradlew.bat
+libs
