--- utils/build.gradle
@@ -1,43 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Java library project to get you started.
- * For more details take a look at the 'Building Java & JVM projects' chapter in the Gradle
- * User Manual available at https://docs.gradle.org/6.7.1/userguide/building_java_projects.html
- */
-
-plugins {
-    // Apply the java-library plugin for API and implementation separation.
-    id 'java-library'
-}
-
-repositories {
-    // Use JCenter for resolving dependencies.
-    jcenter()
-}
-
-dependencies {
-    // Use JUnit test framework.
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.0'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
-
-    // This dependency is exported to consumers, that is to say found on their compile classpath.
-    api 'org.apache.commons:commons-math3:3.6.1'
-
-    // This dependency is used internally, and not exposed to consumers on their own compile classpath.
-    implementation 'com.google.guava:guava:29.0-jre'
-
-    // https://mvnrepository.com/artifact/org.apache.commons/commons-lang3
-     implementation 'org.apache.commons:commons-lang3:3.10'
-
-}
-version = '1.3.5'
-tasks.named('jar') {
-    manifest {
-        attributes('Implementation-Title': project.name,
-                   'Implementation-Version': project.version)
-    }
-}
-test {
-    useJUnitPlatform()
-}
