--- utils/src/main/java/demo/StringUtils.java
@@ -1,21 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package demo;
-
-public class StringUtils {
- private boolean result=false;
-    public boolean isPositiveNumber(String str) {
-       if(str==null){
-         return result;
-       }
-       String newStr=org.apache.commons.lang3.StringUtils.trim(str);
-       result=org.apache.commons.lang3.StringUtils.startsWith(newStr,"-");
-       if(result){
-          return result;
-          } else{
-              result=org.apache.commons.lang3.StringUtils.isNumericSpace(newStr);
-          }
-     return result;
-    }
-}
