--- utils/src/test/java/demo/StringUtilsTest.java
@@ -1,13 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package demo;
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-public class StringUtilsTest {
-    @Test public void testIsPositiveNumber() {
-        StringUtils strUtils = new StringUtils();
-        assertTrue(strUtils.isPositiveNumber("-678"));
-    }
-}
